refactor(discover): move surfaceSectionTypes to types/sectionLayout
Section layout types belong in the types directory alongside
marketDisplayItem, not co-located with components.
