+++ src/features/discover/types/sectionLayout.ts
